requirements/test.txt
bandit==1.8.3
black==25.1.0
flake8==7.1.2
isort==6.0.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-httpx==0.35.0
pytest-mock==3.14.0
pytest-cov==6.1.1
coverage[toml]==7.8.2
pysonar-scanner==0.2.0.520
mypy==1.15.0
mongodb-migrations==1.3.1

asgi-correlation-id==4.3.4
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.0
dependency-injector==4.46.0
pydantic_settings==2.9.1
uuid7==0.1.0
motor==3.7.1
asyncpg==0.30.0
sqlalchemy==2.0.41
greenlet==3.2.2
pyjwt[crypto]==2.10.1
cryptography==45.0.4
redis==6.2.0
git+https://${GITHUB_TOKEN}@github.com/projeto-carreira-luizalabs-2025/pc-logging
